feat(audit-service): added the @PreAuthorize and commented them in code
